Simprede_scrapers/requirements.txt
# Database providers
apache-airflow-providers-postgres==5.8.0
apache-airflow-providers-mysql==5.4.0

# Google Cloud providers
apache-airflow-providers-google==10.11.1
google-cloud-storage==2.10.0
google-cloud-bigquery==3.13.0

# Database drivers
psycopg2-binary>=2.9.9
PyMySQL==1.1.0

# Core dependencies
sqlalchemy==1.4.53
requests>=2.32.0

# Web scraping dependencies
beautifulsoup4==4.12.2
lxml>=4.9.3
lxml_html_clean>=0.1.0
lxml[html_clean]>=4.9.3
selenium==4.21.0

# Data processing
pandas>=2.2.2
numpy>=1.24.4
openpyxl>=3.1.2

# Geospatial processing
geopandas>=0.14.4
shapely>=2.0.4

# NLP basics
nltk>=3.8.1

# Utilities
python-dotenv>=1.0.0
tenacity>=8.2.3
feedparser>=6.0.0
psutil>=5.9.6
urllib3>=1.26.0
python-dateutil>=2.8.0

# Machine Learning
scikit-learn>=1.3.0,<1.5.0
joblib>=1.3.2
imbalanced-learn>=0.11.0

# News processing
newspaper3k>=0.2.8
fake-useragent>=1.4.0

# Async processing
aiohttp>=3.8.0
asyncio-throttle>=1.0.0